requirements/common.txt
Django==1.8.4
Markdown==2.6.2
MySQL-python==1.2.5
Pillow==2.9.0
argparse==1.2.1
boto==2.38.0
django-allauth==0.23.0
django-boto==0.3.9
django-braces==1.8.1
django-extensions==1.5.6
django-filter==0.11.0
django-storages==1.1.8
djangorestframework==3.2.3
djrill==2.2.0
geoip2==0.7.0
gunicorn==19.3.0
ipaddr==2.1.11
maxminddb==1.2.0
oauthlib==1.0.3
python-magic==0.4.6
python-openid==2.2.5
python-social-auth==0.2.12
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
git+https://github.com/mariocesar/sorl-thumbnail@master
sqlparse==0.1.16
wsgiref==0.1.2
PyJWT==1.4.0
python-dateutil==2.4.2
